refactor(text-to-image): moves "/toSharkUIOutput/TextPrompt.ts" to "/toSharkUIOutput/Image/Description/all.ts"
